rest

+++ src/Step/Step7DockerRest.php
@@ -11,10 +11,13 @@ class Step7DockerRest extends Step0Base {
    * Run!
    */
   public function run() : bool {
-    $this->execCommands([
-      "cp -r {$_ENV['ASSETS']}/docker-rest /opt/docker-rest",
-      "/opt/docker-rest/start.sh",
-    ]);
+    if (!file_exists('/opt/docker-rest')) {
+      $this->execCommands([
+        "cp -r {$_ENV['ASSETS']}/docker-rest /opt/docker-rest",
+        "/opt/docker-rest/start.sh",
+      ]);
+    }
+
     return 1;
   }
